Move LiveStreamController to be under Coronavirus namespace
This puts this code under the Coronavirus namespace to reflect that this
is just for the Coronavirus code of this app.
