+++ app/views/coronavirus/coronavirus_pages/index.html.erb
@@ -3,7 +3,7 @@
 <h2 class="govuk-heading-s govuk-!-margin-bottom-2">Coronavirus landing page</h2>
 <ul class="govuk-list">
   <li><%= link_to t("coronavirus_pages.index.landing_page_edit.accordions"), coronavirus_page_path(slug: @topic_page.slug), class:"govuk-link" %></li>
-  <li><%= link_to t("coronavirus_pages.index.landing_page_edit.live_stream_url"), live_stream_index_path, class:"govuk-link" %></li>
+  <li><%= link_to t("coronavirus_pages.index.landing_page_edit.live_stream_url"), coronavirus_live_stream_index_path, class:"govuk-link" %></li>
   <li><%= link_to t("coronavirus_pages.index.landing_page_edit.something_else"), prepare_coronavirus_page_path(slug: @topic_page[:slug]), class:"govuk-link" %></li>
 </ul>
